Move engine types to types.py, default use_topk_divergence=True
- Extract SampleCore, ScalarPreparedSample, TopKPreparedSample,
  ScalarBehavior, TopKBehavior into claas/training/engine/tinker/types.py
- Change use_topk_divergence default from False to True in TrainingConfig
  and base.yaml — top-K GJS is now the default divergence mode
