+++ claas/core/types.py
@@ -32,7 +32,7 @@ class TrainingConfig:
     max_grad_norm: float = 1.0
     kl_reg_weight: float = 0.0
     teacher_top_k: int = 100
-    use_topk_divergence: bool = False
+    use_topk_divergence: bool = True
     steps_per_batch: int = 4
     feedback_repetitions: int = 1
