+++ claas/eval/configs/base.yaml
@@ -35,6 +35,6 @@ training:
   max_grad_norm: 1.0
   kl_reg_weight: 0.0
   teacher_top_k: 100
-  use_topk_divergence: false
+  use_topk_divergence: true
   steps_per_batch: 4
   feedback_repetitions: 1
